+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Create the `Mosaic` component
     -   Create the `Tabs` component family (`Tab`, `TabList`, `TabPanel`, `TabProvider`)
     -   Create the `ImageBlock` component
+    -   Create the `Combobox` component family
 -   `@lumx/react`:
     -   Create the `Combobox` component family